Fix some warnings

+++ libnemo-private/nemo-canvas-item.c
@@ -1601,7 +1601,7 @@ nemo_canvas_item_event (EelCanvasItem *item, GdkEvent *event)
 	canvas_item = NEMO_CANVAS_ITEM (item);
 	cursor_window = ((GdkEventAny *)event)->window;
 
-	switch (event->type) {
+	switch ((int) event->type) {
 	case GDK_ENTER_NOTIFY:
         gtk_widget_set_tooltip_text (GTK_WIDGET (item->canvas), canvas_item->tooltip);
 		if (!canvas_item->details->is_prelit) {
